requirements.txt
pyaudio
numpy
soundfile
sounddevice
webrtcvad
ollama
piper-tts
simpleaudio
faster-whisper
noisereduce